requirements.txt
Django==3.2.4
django-cors-headers==3.7.0
django-extensions==2.2.9
dj-rest-auth==1.0.4
djangorestframework==3.12.4
django-ckc==0.0.5

# Heroku staging debug tools
django-debug-toolbar==3.2.1
django-querycount==0.7.0

# Database
dj-database-url==0.5.0
psycopg2-binary==2.8.6

# Storage
whitenoise==5.1.0

# Email
django-anymail==8.1
django-templated-email==2.3.0

# Monitoring
scout-apm==2.19.0

# realtime
channels==3.0.3
uvicorn[standard]==0.14.0
Twisted==20.3.0
channels_redis==3.2.0
service-identity==21.1.0
